+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.13 (unreleased)
+## 0.0.13 (2017-06-28)
 
 -   Change FAR threshold for creation of superevents to 1 per day.